dec_2023/Vaibhav_Gupta/requirements.txt
altair==5.2.0
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
bokeh==2.4.3
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.2
idna==3.6
importlib-metadata==6.11.0
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
numpy==1.26.2
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
plotly==5.18.0
protobuf==4.25.1
pyarrow==14.0.1
pydeck==0.8.1b0
Pygments==2.17.2
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
rich==13.7.0
rpds-py==0.13.2
six==1.16.0
smmap==5.0.1
soupsieve==2.5
SQLAlchemy==1.4.0
streamlit==1.29.0
streamlit-nested-layout==0.1.1
streamlit-vis-timeline==0.3.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.4
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
xyzservices==2023.10.1
zipp==3.17.0

